Editor Update
Tiny MCE has been implemented.

I have been able to get the editor working and post content to the server.
But when it is retrieved it is not correctly displayed.

Note: Tiny MCE is implemented as a controlled component. This means I am updating the state which each key stroke. Which could prove problematic in the performance area for larger documents. This will have to stay the current implementation for now.