News Section
How data is passed and retrieved has been updated.
Crypto list is called by News component and passed to the NewsList. The NewsList will check for the crypto list and build a list of ticker symbols and dispatch the action along with the page: getNewsArticles(cryptoList, page). Finally, it will create a news card for each news article that was sent from the backend.